Python2.5 compatibility
