+++ scikits/learn/manifold/tests/test_locally_linear.py
@@ -1,15 +1,15 @@
-import itertools
 import numpy as np
 
 from numpy.testing import assert_array_almost_equal
 from scikits.learn import neighbors, manifold
+from scikits.learn.utils.fixes import product
 
 #----------------------------------------------------------------------
 # Test LLE by computing the reconstruction error on some manifolds.
 
 def test_lle_simple_grid():
     # grid of equidistant points in 2D, out_dim = n_dim
-    X = np.array(list(itertools.product(range(5), repeat=2)))
+    X = np.array(list(product(range(5), repeat=2)))
     clf = manifold.LocallyLinearEmbedding(n_neighbors=5, out_dim=2)
     tol = .1
 
@@ -29,7 +29,7 @@ def test_lle_simple_grid():
 
 def test_lle_manifold():
     # similar test on a slightly more complex manifold
-    X = np.array(list(itertools.product(range(20), repeat=2)))
+    X = np.array(list(product(range(20), repeat=2)))
     X = np.c_[X, X[:, 0]**2 / 20]
     clf = manifold.LocallyLinearEmbedding(n_neighbors=5, out_dim=2)
     tol = .5
